Address Chris' review comments.
Make the siteId in the LargeTempTableBlockId be 64 bits.  This means we
can't have a single 64 bit integer for it.  But we need to pass it and
the block id up to Java through the JNI interface.  So we still need it
to be unsigned.

Change the Java LargeTempTableBlockManager to keep structured keys for
blocks rather than simple integers.  We could have done this with Java's
BigInteger and some arithmetic, but that seems cluttered.

+++ src/ee/common/LargeTempTableBlockCache.cpp
@@ -39,7 +39,7 @@ LargeTempTableBlockCache::LargeTempTableBlockCache(Topend *topend, int64_t maxCa
     , m_totalAllocatedBytes(0)
 {
     ExecutorContext *ec = ExecutorContext::getExecutorContext();
-    int32_t siteId = ec ? ec->m_siteId : 0;
+    int64_t siteId = ec ? ec->m_siteId : 0;
     m_nextId = LargeTempTableBlockId(siteId, 0);
 }
